all: Add bitwise operations
Same syntax as C, but without the assignment-by versions and with
the addition of <<< (rotate left) and >>> (rotate right).

Disable constant folding in get_binary_expression_type() because it
was interfering with binary operand verification. This means that
"x := 0xffffffff + 1" resolves to u32 again, but that is subject to
change.