base: Add File::view()
If we assign to a string_view from File::map, it will use strlen() to
get the length which breaks if the file size is a multiple of the page
size as there will be no zeroes added to the end of the page, so we will
keep reading past the end. Fix with a helper function.

+ In make_file_mappable(), only re-extend files opened with the
  TRUNCATE flag if the READ flag is also set.
+ Unify some shared code.